refactor(core-api): add schema for orderBy query param (#3243)
