+++ packages/core-api/src/handlers/shared/schemas/index.ts
@@ -1,0 +1,6 @@
+import { blockId } from "./block-id";
+import { orderBy } from "./order-by";
+import { pagination } from "./pagination";
+import { walletId } from "./wallet-id";
+
+export { blockId, orderBy, pagination, walletId };
