+++ packages/core-api/src/handlers/transactions/schema.ts
@@ -1,7 +1,6 @@
 import { app } from "@arkecosystem/core-container";
 import Joi from "@hapi/joi";
-import { blockId } from "../shared/schemas/block-id";
-import { pagination } from "../shared/schemas/pagination";
+import { blockId, orderBy, pagination } from "../shared/schemas";
 
 const address: object = Joi.string()
     .alphanum()
@@ -11,7 +10,7 @@ export const index: object = {
     query: {
         ...pagination,
         ...{
-            orderBy: Joi.string(),
+            orderBy,
             id: Joi.string()
                 .hex()
                 .length(64),
@@ -101,7 +100,7 @@ export const search: object = {
         },
     },
     payload: {
-        orderBy: Joi.string(),
+        orderBy,
         id: Joi.string()
             .hex()
             .length(64),
